# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for fleet management",
        "title": "Fleet Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/inventory": {
            "get": {
                "description": "Trả về toàn bộ vật tư trong kho",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Lấy danh sách tồn kho",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Inventory"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Tạo mới vật tư",
                "parameters": [
                    {
                        "description": "Inventory body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Inventory"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Inventory"
                        }
                    }
                }
            }
        },
        "/inventory-log": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "InventoryLog"
                ],
                "summary": "Get all inventory logs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.InventoryLog"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "InventoryLog"
                ],
                "summary": "Create inventory log",
                "parameters": [
                    {
                        "description": "InventoryLog",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.InventoryLog"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.InventoryLog"
                        }
                    }
                }
            }
        },
        "/inventory-log/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "InventoryLog"
                ],
                "summary": "Get one log",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.InventoryLog"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "InventoryLog"
                ],
                "summary": "Update inventory log",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "InventoryLog",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.InventoryLog"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.InventoryLog"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "InventoryLog"
                ],
                "summary": "Delete inventory log",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/inventory/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Lấy chi tiết vật tư",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Inventory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Inventory"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Cập nhật vật tư",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Inventory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Inventory body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Inventory"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Inventory"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Xóa vật tư",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Inventory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Inventory": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "Ghi chú",
                    "type": "string"
                },
                "itemCode": {
                    "description": "Mã vật tư",
                    "type": "string"
                },
                "itemName": {
                    "description": "Tên vật tư",
                    "type": "string"
                },
                "minStock": {
                    "description": "Mức tồn kho cảnh báo",
                    "type": "integer"
                },
                "quantity": {
                    "description": "Tồn kho",
                    "type": "integer"
                },
                "unit": {
                    "description": "Đơn vị tính",
                    "type": "string"
                },
                "warehouse": {
                    "description": "Tên kho",
                    "type": "string"
                }
            }
        },
        "models.InventoryLog": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "IN | OUT | ADJUST",
                    "type": "string"
                },
                "inventory_id": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}